requirements.txt
Flask
redis
python-dotenv
gunicorn
qdrant-client
sentence-transformers
tiktoken
arxiv
docling
